# Add ToStringBuilder helper to support uniform ToString format