--- a/src/Elastic.Apm/Helpers/ToStringBuilder.cs
+++ b/src/Elastic.Apm/Helpers/ToStringBuilder.cs
@@ -1,32 +1,37 @@
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 
 namespace Elastic.Apm.Helpers
 {
-	public class ToStringBuilder
+	//
+	// We need to implement IEnumerable<string> to support collection initializer syntax
+	//
+	public struct ToStringBuilder : IEnumerable<string>
 	{
-		private readonly StringBuilder _stringBuilder = new StringBuilder();
+		private const int _stringBuilderInitialCapacity = 100;
+		private readonly StringBuilder _stringBuilder;
+		private bool _addedAny;
 
 		public ToStringBuilder(string className)
 		{
+			_stringBuilder = new StringBuilder(_stringBuilderInitialCapacity);
 			_stringBuilder.Append(className);
 			_stringBuilder.Append("{");
+			_addedAny = false;
 		}
 
-		private bool _addedAny;
-
-		public object this[string propertyName]
-		{
-			set => Add(propertyName, value);
-		}
-
-		private void Add(string propertyName, object propertyValue)
+		public void Add<T>(string propertyName, T propertyValue)
 		{
 			if (_addedAny) _stringBuilder.Append(", ");
-
 			_stringBuilder.Append(propertyName).Append(": ").Append(propertyValue);
-
 			_addedAny = true;
 		}
+
+		public IEnumerator<string> GetEnumerator() => Enumerable.Empty<string>().GetEnumerator();
+
+		IEnumerator IEnumerable.GetEnumerator() => Enumerable.Empty<string>().GetEnumerator();
 
 		public override string ToString() => _stringBuilder + "}";
 	}