--- a/src/Elastic.Apm/Helpers/ToStringBuilder.cs
+++ b/src/Elastic.Apm/Helpers/ToStringBuilder.cs
@@ -1,14 +1,14 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 
 namespace Elastic.Apm.Helpers
 {
 	//
-	// We need to implement IEnumerable<string> to support collection initializer syntax
+	// We need to implement IEnumerable to support collection initializer syntax
 	//
-	public struct ToStringBuilder : IEnumerable<string>
+	public struct ToStringBuilder : IEnumerable
 	{
 		private const int _stringBuilderInitialCapacity = 100;
 		private readonly StringBuilder _stringBuilder;
@@ -21,16 +21,14 @@ namespace Elastic.Apm.Helpers
 			_addedAny = false;
 		}
 
-		public void Add(string propertyName, object propertyValue)
+		public void Add<T>(string propertyName, T propertyValue)
 		{
 			if (_addedAny) _stringBuilder.Append(", ");
 			_stringBuilder.Append(propertyName).Append(": ").Append(propertyValue);
 			_addedAny = true;
 		}
 
-		public IEnumerator<string> GetEnumerator() => Enumerable.Empty<string>().GetEnumerator();
-
-		IEnumerator IEnumerable.GetEnumerator() => Enumerable.Empty<string>().GetEnumerator();
+		IEnumerator IEnumerable.GetEnumerator() => throw new NotImplementedException();
 
 		public override string ToString() => _stringBuilder + "}";
 	}