--- a/src/Elastic.Apm/Model/Payload/Error.cs
+++ b/src/Elastic.Apm/Model/Payload/Error.cs
@@ -43,6 +43,12 @@ namespace Elastic.Apm.Model.Payload
 		[JsonProperty("transaction_id")]
 		public string TransactionId { get; set; }
 
-		public override string ToString() => $"Error, Id: {Id}, TraceId: {TraceId}, ParentId: {ParentId}, TransactionId: {TransactionId}";
+		public override string ToString() => new ToStringBuilder(nameof(Error))
+		{
+			["Id"] = Id,
+			["TraceId"] = TraceId,
+			["ParentId"] = ParentId,
+			["TransactionId"] = TransactionId
+		}.ToString();
 	}
 }