--- a/test/Elastic.Apm.Tests/Elastic.Apm.Tests.csproj
+++ b/test/Elastic.Apm.Tests/Elastic.Apm.Tests.csproj
@@ -24,6 +24,7 @@
     <ProjectReference Include="..\..\src\Elastic.Apm\Elastic.Apm.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="HelpersTests" />
     <Folder Include="Mocks\" />
     <Folder Include="ApiTests\" />
   </ItemGroup>