Added ToStringBuilder helper to support uniform ToString format

- Implemented ToStringBuilder helper
- Added tests to cover ToStringBuilder helper
- Replaced existing ToString ad-hoc implementations with ones using ToStringBuilder