@gregkalapos 2019-03-21
Measured the difference between this and the current logging code on `master`. 

Test case (both ran on `master` and on the `pr branch`):
- 1. test: create 100transaction and 10span on each transaction and end those without any logging (loglevel set to `Error`, no log in this case)
- 2. test: same as 1. test, but with loglevel `Debug`, which utilizes the part of the code with the pr. 

`master` performance: 
```
|                              Method |      Mean |     Error |    StdDev |    Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------ |----------:|----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|           SimpleTransactions10Spans |  1.153 ms | 0.0157 ms | 0.0147 ms |  1.150 ms |    228.5156 |    113.2813 |           - |             1.37 MB |
| DebugLogSimple100Transaction10Spans | 24.687 ms | 2.8610 ms | 8.4359 ms | 27.703 ms |   1656.2500 |    531.2500 |    125.0000 |             6.43 MB |
```

PR performance: 
```
|                              Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------ |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|           SimpleTransactions10Spans |  1.100 ms | 0.0111 ms | 0.0092 ms |    228.5156 |    113.2813 |           - |             1.37 MB |
| DebugLogSimple100Transaction10Spans | 26.349 ms | 2.5303 ms | 7.4607 ms |   1468.7500 |    515.6250 |    140.6250 |             8.74 MB |
```