@SergeyKleyman 2019-03-27
I ran the benchmark project added by @gregkalapos on my machine (Windows 10 on MacBook Pro - Intel Core i9-8950HK CPU 2.90GHz, 1 CPU, 12 logical and 6 physical cores).

Results based on code in master on my machine:
```
|                              Method |        Mean |       Error |       StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------ |------------:|------------:|-------------:|------------:|------------:|------------:|--------------------:|
|           SimpleTransactions10Spans |    933.9 us |    18.34 us |     21.12 us |    228.5156 |    114.2578 |           - |             1.37 MB |
| DebugLogSimple100Transaction10Spans | 40,771.0 us | 3,499.04 us | 10,095.52 us |   1285.7143 |    500.0000 |           - |             6.43 MB |
```

Results based on code in this PR:
```
|                              Method |        Mean |       Error |      StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------ |------------:|------------:|------------:|------------:|------------:|------------:|--------------------:|
|           SimpleTransactions10Spans |    905.1 us |    13.65 us |    12.77 us |    228.5156 |    114.2578 |           - |             1.37 MB |
| DebugLogSimple100Transaction10Spans | 29,710.7 us | 3,318.01 us | 9,519.99 us |   1562.5000 |    500.0000 |           - |             8.05 MB |
```


Results based on code in this PR except for ToStringBuilder.Add method being generic:
```
|                              Method |        Mean |       Error |      StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------ |------------:|------------:|------------:|------------:|------------:|------------:|--------------------:|
|           SimpleTransactions10Spans |    912.4 us |    11.65 us |    10.90 us |    228.5156 |    114.2578 |           - |             1.37 MB |
| DebugLogSimple100Transaction10Spans | 31,196.1 us | 2,769.31 us | 8,078.21 us |   1625.0000 |    593.7500 |           - |             8.05 MB |
```